Add carrnexa namespace and rename package to app_name
