+++ src/carrnexa/app_name/__main__.py
@@ -1,0 +1,4 @@
+from carrnexa.app_name.cli.main import app
+
+if __name__ == "__main__":
+    app()
