--- src/python_template/__main__.py
@@ -1,4 +1,0 @@
-from python_template.cli.main import app
-
-if __name__ == "__main__":
-    app()
